+++ HolaConsultores.OnlineShop.Infraestructure/Repositories/ProductColor/ProductColorRepository.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using HolaConsultores.OnlineShop.Domain.Interfaces.IRepositories;
 using HolaConsultores.OnlineShop.Infraestructure.Context;
-using HolaConsultores.OnlineShop.Infraestructure.Entities;
+using HolaConsultores.OnlineShop.Infraestructure.Entities.ProductColor;
 using Microsoft.EntityFrameworkCore;
 
 namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductColor
@@ -16,20 +16,20 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductColor
         {
         }
 
-        #region GET (Not implemented)
-        public Task<IEnumerable<ProductColorModel>> GetAllAsync()
+        #region GET
+        public async Task<IEnumerable<ProductColorModel>> GetAllAsync()
         {
-            throw new NotImplementedException();
+            return await _context.ProductColors.ToListAsync();
         }
 
-        public Task<IEnumerable<ProductColorModel>> GetAllAsync(int offset, int limit)
+        public async Task<IEnumerable<ProductColorModel>> GetAllAsync(int offset, int limit)
         {
-            throw new NotImplementedException();
+            return await _context.ProductColors.Skip(offset).Take(limit).ToListAsync();
         }
 
-        public Task<ProductColorModel> GetByIdAsync(int id)
+        public async Task<ProductColorModel> GetByIdAsync(int id)
         {
-            throw new NotImplementedException();
+            return await _context.ProductColors.Where(pC => pC.Id == id).FirstOrDefaultAsync();
         }
         #endregion
 
@@ -38,9 +38,21 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductColor
         {
             try
             {
-                var result = await _context.ProductColors.AddAsync(obj);
-                _context.SaveChanges();
-                return result.Entity;
+               
+
+                var productColors = await GetAllAsync();
+
+                if (!productColors.Any(x => x.ProductId == obj.ProductId && x.ColorId == obj.ColorId))
+                {
+                    var result = await _context.ProductColors.AddAsync(obj);
+                    _context.SaveChanges();
+                    return result.Entity;
+                }
+                else
+                {
+                    throw new Exception("This product already have this color.");
+                }
+
             }
             catch (Exception ex)
             {
@@ -55,9 +67,18 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductColor
             try
             {
                 var result = await GetByIdAsync(id);
-                _context.ProductColors.Remove(result);
-                await _context.SaveChangesAsync();
-                return result;
+               
+
+                if (result != null)
+                {
+                    _context.ProductColors.Remove(result);
+                    await _context.SaveChangesAsync();
+                    return result;
+                }
+                else
+                {
+                    throw new Exception("There are not any productColor with the specified id");
+                }
             }
             catch (Exception ex)
             {
