+++ HolaConsultores.OnlineShop.Infraestructure/Repositories/ProductSize/ProductSizeRepository.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using HolaConsultores.OnlineShop.Domain.Interfaces.IRepositories;
 using HolaConsultores.OnlineShop.Infraestructure.Context;
-using HolaConsultores.OnlineShop.Infraestructure.Entities;
+using HolaConsultores.OnlineShop.Infraestructure.Entities.ProductSize;
 using Microsoft.EntityFrameworkCore;
 
 namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductSize
@@ -16,20 +16,20 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductSize
         {
         }
 
-        #region GET (Not implemented)
-        public Task<IEnumerable<ProductSizeModel>> GetAllAsync()
+        #region GET
+        public async Task<IEnumerable<ProductSizeModel>> GetAllAsync()
         {
-            throw new NotImplementedException();
+            return await _context.ProductSizes.ToListAsync();
         }
 
-        public Task<IEnumerable<ProductSizeModel>> GetAllAsync(int offset, int limit)
+        public async Task<IEnumerable<ProductSizeModel>> GetAllAsync(int offset, int limit)
         {
-            throw new NotImplementedException();
+            return await _context.ProductSizes.Skip(offset).Take(limit).ToListAsync();
         }
 
-        public Task<ProductSizeModel> GetByIdAsync(int id)
+        public async Task<ProductSizeModel> GetByIdAsync(int id)
         {
-            throw new NotImplementedException();
+            return await _context.ProductSizes.Where(pC => pC.Id == id).FirstOrDefaultAsync();
         }
         #endregion
 
@@ -38,9 +38,18 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductSize
         {
             try
             {
-                var result = await _context.ProductSizes.AddAsync(obj);
-                _context.SaveChanges();
-                return result.Entity;
+                var productSize = await GetAllAsync();
+
+                if (!productSize.Any(x => x.ProductId == obj.ProductId && x.SizeId == obj.SizeId))
+                {
+                    var result = await _context.ProductSizes.AddAsync(obj);
+                    _context.SaveChanges();
+                    return result.Entity;
+                }
+                else
+                {
+                    throw new Exception("This product already have this size.");
+                }
             }
             catch (Exception ex)
             {
@@ -55,9 +64,17 @@ namespace HolaConsultores.OnlineShop.Infraestructure.Repositories.ProductSize
             try
             {
                 var result = await GetByIdAsync(id);
-                _context.ProductSizes.Remove(result);
-                await _context.SaveChangesAsync();
-                return result;
+
+                if (result != null)
+                {
+                    _context.ProductSizes.Remove(result);
+                    await _context.SaveChangesAsync();
+                    return result;
+                }
+                else
+                {
+                    throw new Exception("There are not any productSize with the specified id");
+                }
             }
             catch (Exception ex)
             {
